requirements.txt
python-dotenv~=1.0.0
selenium~=3.14.1
pytest~=7.0.0